Put the file in the wrong directory.
