+++ playitagainsam/recorder.py
@@ -77,21 +77,35 @@ class Recorder(SocketCoordinator):
             os.close(proc_fd)
         super(Recorder, self).cleanup()
 
+    def _utf8_length_first_byte(self, byte):
+        # https://en.wikipedia.org/wiki/UTF-8#Description
+        byte_int = ord(byte)
+        if byte_int & 0b11111100 == 0b11111100:
+            return 6
+        elif byte_int & 0b11111000 == 0b11111000:
+            return 5
+        elif byte_int & 0b11110000 == 0b11110000:
+            return 4
+        elif byte_int & 0b11100000 == 0b11100000:
+            return 3
+        elif byte_int & 0b11000000 == 0b11000000:
+            return 2
+        elif byte_int & 0b10000000 == 0b10000000:
+            # 0b10 in the significant bits is a flag for continuation bytes. We
+            # should not see this on a utf8 character start. This is not good.
+            raise OSError
+        else:
+            return 1
+
+    def _get_utf8_character(self, view_fd):
+        partial_input = [self._read_one_byte(view_fd)]
+        for _ in xrange(self._utf8_length_first_byte(partial_input[0]) - 1):
+            partial_input.append(self._read_one_byte(view_fd))
+        return "".join(partial_input)
+
     def _handle_input(self, view_fd):
         try:
-            # We assume all I/O is in utf8, so we need to read
-            # a complete utf8 character before doing anything else.
-            # This might be encoded into multiple bytes of input.
-            c = None
-            input = self._read_one_byte(view_fd)
-            while c is None:
-                try:
-                    c = input.decode("utf8")
-                except UnicodeDecodeError:
-                    # No valid utf8 sequence is longer than 6 bytes.
-                    if len(input) >= 6:
-                        raise
-                    input += self._read_one_byte(view_fd)
+            c = self._get_utf8_character(view_fd)
         except OSError:
             pass
         else:
@@ -104,16 +118,7 @@ class Recorder(SocketCoordinator):
                 "data": c,
             })
             # Forward it to the corresponding terminal process.
-            os.write(proc_fd, input)
-
-    def _resize_event_thread(self):
-        # Check if we have a resize event.
-        for _ in iter(self.resize_event.get, None):
-            size = get_terminal_size(1)
-            for term in self.terminals:
-                _unused_1, proc_fd, _unused_2 = self.terminals[term]
-                set_terminal_size(proc_fd, size)
-
+            os.write(proc_fd, c)
 
     def _handle_output(self):
         ready = self.wait_for_data(self.proc_fds, 0.01)
@@ -127,9 +132,13 @@ class Recorder(SocketCoordinator):
             # because multiple bytes might be part of a single utf8 char.
             proc_output = []
             proc_ready = [proc_fd]
-            while proc_ready:
+            character_count = 0
+            while proc_ready and character_count < 4096:
+                # Loop until we have no new data or we have >= 4096 characters
+                # processed.
                 try:
-                    c = self._read_one_byte(proc_fd)
+                    c = self._get_utf8_character(proc_fd)
+                    character_count += 1
                 except OSError:
                     if proc_output:
                         self.eventlog.write_event({
@@ -149,8 +158,16 @@ class Recorder(SocketCoordinator):
                 self.eventlog.write_event({
                     "act": "WRITE",
                     "term": term,
-                    "data": six.b("").join(proc_output).decode("utf8"),
+                    "data": six.b("").join(proc_output),
                 })
+
+    def _resize_event_thread(self):
+        # Check if we have a resize event.
+        for _ in iter(self.resize_event.get, None):
+            size = get_terminal_size(1)
+            for term in self.terminals:
+                _unused_1, proc_fd, _unused_2 = self.terminals[term]
+                set_terminal_size(proc_fd, size)
 
     def _read_one_byte(self, fd):
         """Read a single byte, or raise OSError on failure."""
